Fix/update builds for Python 3.7 and 3.8
Remove calls to deprecated functions in setup.py:
    - platform.linux_distribution() to platform.platform()
    - setuptools.dist._get_unpatched() to
      setuptools.monkey.get_unpatched()

Fix URL for freetds in build_manylinux_wheels. The "stable" link
is not working, use fixed version 1.1.

Change manylinux container from manylinux1 to manylinux2010.

Limit manylinux wheel builds and testing to Python 3.7 and 3.8,
which are the current defaults in Debian Stable and Archlinux,
respectively.